What I have so far
This is what I have so far, but it is all needs work. The last two are what I have made for the web page. The css is designed to run in a folder called "css". I also made the html so that a javascript file connects to it, although I'm not sure we need it, as I don't know what methods we will use to implement our java here. 

The other two are for the main program. I named them with my initials, so they are not confused with other work. Hope they help!